backend/requirements.txt
async-timeout==4.0.3
annotated-types==0.6.0
starlette==0.35.1
typing_extensions==4.9.0
fastapi==0.109.0
uvicorn[standard]==0.23.2
sqlalchemy[asyncio] >= 2.0
pylint>=3.0
asyncpg >=0.29
python-dotenv>=1.0
strawberry-graphql[fastapi] == 0.219.1
strawberry-sqlalchemy-mapper == 0.3.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
uvicorn==0.23.2

